Refactoring for client. Added comments, docs, some redesign. Remove some mistake from server part.

+++ Client/internal/requseter/deleteData.go
@@ -8,44 +8,51 @@ import (
 	"os"
 	"strings"
 
+	"github.com/Arti9991/GoKeeper/client/internal/clientmodels"
 	pb "github.com/Arti9991/GoKeeper/client/internal/proto"
 	"google.golang.org/grpc"
 	"google.golang.org/grpc/metadata"
 )
 
-func DeleteDataRequest(StorageID string, req *ReqStruct, offlineMode bool) error {
-
+// DeleteDataRequest метод удаления данных
+func (req *ReqStruct) DeleteDataRequest(StorageID string, offlineMode bool) error {
 	var err error
-
-	fmt.Println(offlineMode)
+	// спрашиваем подтверждение на выход из аккаунта
+	fmt.Println("Вы точно хотите удалить данные? [y/n] (В онлайн режиме данные на сервере будут также удалены!)")
+	reader := bufio.NewReader(os.Stdin)
+	// читаем путь из консоли
+	agree, err := reader.ReadString('\n')
+	agree = strings.ToLower(agree)
+	if strings.Contains(agree, "n") {
+		return clientmodels.ErrUserAbort
+	}
 
 	if !offlineMode {
+		// если режим работы онлайн, то удаляем данные на сервере
 		err = DeleteDataOnline(StorageID, req.ServAddr, req)
 		if err != nil {
-			fmt.Println(err)
+			fmt.Printf("\nНе удалось удалить данные на сервере: %s\n", err.Error())
 		}
 	}
-
+	// удаляем информацию о данных из локальной базы
 	err = req.DBStor.DeleteData(StorageID)
 	if err != nil {
-		fmt.Println(err)
 		return err
 	}
+	// удаляем сами данные из хранилища
 	err = req.BinStor.RemoveBinData(StorageID)
 	if err != nil {
-		fmt.Println(err)
 		return err
 	}
 	return nil
 }
 
+// DeleteDataOnline функция удаления данных с сервера
 func DeleteDataOnline(StorageID string, addr string, req *ReqStruct) error {
 
-	fmt.Println("Open token")
+	// считываем токен из файла
 	file, err := os.Open("./Token.txt")
 	if err != nil {
-		fmt.Println(err)
-		//logger.Log.Error("SAVE Error in opening file", zap.Error(err))
 		return err
 	}
 	defer file.Close()
@@ -54,28 +61,26 @@ func DeleteDataOnline(StorageID string, addr string, req *ReqStruct) error {
 	// Считываем строку текста
 	UserID, err := reader.ReadString('\n')
 	if err != nil {
-		fmt.Println(err)
 		return err
 	}
-	//Выводим строку
+	// удаляем суффикс
 	UserID = strings.TrimSuffix(UserID, "\n")
-	fmt.Printf("%#v", UserID)
 
+	// добавляем метаданные с UserID к запросу
 	var header metadata.MD
 	md := metadata.New(map[string]string{"UserID": UserID})
 	ctx := metadata.NewOutgoingContext(context.Background(), md)
-
+	// инциализируем клиент
 	dial, err := grpc.NewClient(addr, grpc.WithTransportCredentials(req.Creds)) //req.ServAddr
 	if err != nil {
 		log.Fatal(err)
 	}
-
+	// вызываем метод удаления на сервере
 	r := pb.NewKeeperClient(dial)
 	_, err = r.DeleteData(ctx, &pb.DeleteDataRequest{
 		StorageID: StorageID,
 	}, grpc.Header(&header))
 	if err != nil {
-		fmt.Println(err)
 		return err
 	}
 	return nil
